style: 피드백 수정:atom 타입 위치 치 ㄱ
