+++ just-plan/src/store/auth.atom.type.ts
@@ -1,0 +1,12 @@
+export interface UserInfo {
+  email: string;
+  id: number | null;
+  name: string;
+  isLoggedIn: boolean;
+}
+export const initialUserInfo: UserInfo = {
+  email: "",
+  id: null,
+  name: "",
+  isLoggedIn: false,
+};
